apps/api/pyproject.toml
[project]
name = "examiner-api"
version = "0.1.0"
description = "AI Examiner API - FastAPI backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "redis>=5.2.0",
    "boto3>=1.35.0",
    "anthropic>=0.40.0",
    "groq>=0.12.0",
    "authlib>=1.4.0",
    "httpx>=0.28.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.0,<5.0.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.18",
    "tenacity>=9.0.0",
    "pybreaker>=1.2.0",
    "celery[redis]>=5.4.0",
    "pgvector>=0.3.6",
    "pymupdf>=1.25.0",
    "sentence-transformers>=3.3.0",
    "langdetect>=1.0.9",
    "jinja2>=3.1.0",
    "email-validator>=2.0.0",
    "python-docx>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "factory-boy>=3.3.0",
    "vcrpy>=7.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "types-passlib>=1.7.0",
    "types-python-jose>=3.3.0",
    "types-redis>=4.6.0",
    "boto3-stubs>=1.35.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM", "TCH", "RUF"]
ignore = ["A003", "B008", "TC001", "TC002", "TC003", "RUF001", "RUF003"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["authlib.*", "langdetect.*", "jose.*", "passlib.*", "boto3.*", "botocore.*", "fitz.*", "docx.*", "groq.*", "anthropic.*", "sentence_transformers.*", "pybreaker.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
